pyproject.toml
[tool.poetry]
name = "ultrastar-server"
version = "0.1.0"
description = ""
authors = ["pheelwell"]
readme = "README.md"
#packages = [{include = "ultrastar_server"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
sqlalchemy = "1.4.18"
flask = "2.2.0"
flask-sqlalchemy = "3.0.0"
pyinstaller = "^5.11.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
run-server = "flask:main.debug.run(host='0.0.0.0')"
index-db = "index:index_songs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
env_files = [
    ".env",
    ".test.env"]
testpaths = [
    "test",
    "integration",
]